test(mitm-addon): split requestheaders tests by contract
